+++ jpa-spec-kotlin-dsl-hibernate/src/main/kotlin/io/github/alfonsoristorato/jpaspeckotlindslhibernate/internal/ExperimentalHibernateApi.kt
@@ -1,4 +1,4 @@
-package io.github.alfonsoristorato.jpaspeckotlindsl.internal
+package io.github.alfonsoristorato.jpaspeckotlindslhibernate.internal
 
 @RequiresOptIn(
     message =
